refactor(detector-aws): add Beanstalk detector implementing DetectorSync interface
